Move the ansi related code to utils
